+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v.0.20.0 (2022-06-28)
+
+- changes client call signature to enable privkey hydration
+- adds privkey caching to improve performance
+
 # v.0.19.0 (2022-06-27)
 
 - decreases build size and updates peerDependencies